# packages/api-contract/openapi/footalert.v1.yaml
openapi: 3.1.0
info:
  title: FootAlert BFF API
  version: 1.0.0
  description: |
    Public contract consumed by FootAlert clients (mobile, web, desktop).
    This spec intentionally uses broad payload schemas because upstream API-Football
    structures can evolve, while preserving endpoint/query/path contracts.
servers:
  - url: http://localhost:3001
    description: Local BFF server
  - url: https://staging-api.footalert.example.com
    description: Staging BFF server
paths:
...
  /v1/mobile/session/challenge:
    post:
      summary: Create short-lived mobile attestation challenge
      operationId: createMobileSessionChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileSessionChallengeRequest'
      responses:
        '200':
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileSessionChallengeResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
  /v1/mobile/session/attest:
    post:
      summary: Verify mobile attestation and issue short-lived session token
      operationId: attestMobileSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileSessionAttestRequest'
      responses:
        '200':
          description: Session token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileSessionAttestResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
  /v1/notifications/tokens:
    post:
      summary: Register push token (attested mobile session)
      operationId: registerPushToken
      security:
        - MobileSessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushTokenPayload'
      responses:
        '200':
          description: Token registered
          content:
            application/json:
              schema:
                type: object
                required: [ status, token ]
                properties:
                  status:
                    type: string
                    enum: [ registered ]
                  token:
                    type: string
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
  /v1/notifications/tokens/{token}:
    delete:
      summary: Revoke push token (attested mobile session)
      operationId: revokePushToken
      security:
        - MobileSessionBearer: []
      parameters:
        - $ref: '#/components/parameters/TokenPath'
      responses:
        '204':
          description: Token revoked
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
...
components:
  securitySchemes:
    MobileSessionBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TokenPath:
      name: token
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
...
    ProblemResponse:
      description: Error payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemResponse'
  schemas:
...
    ProblemResponse:
      type: object
      required: [ error, message ]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: 'null'
    MobileSessionChallengeRequest:
      type: object
      additionalProperties: false
      required: [ platform, deviceIdHash, appVersion, buildNumber ]
      properties:
        platform:
          type: string
          enum: [ android, ios ]
        deviceIdHash:
          type: string
          minLength: 8
          maxLength: 256
        appVersion:
          type: string
          minLength: 1
          maxLength: 64
        buildNumber:
          type: string
          minLength: 1
          maxLength: 64
    MobileSessionChallengeResponse:
      type: object
      additionalProperties: false
      required: [ challengeId, challenge, expiresAtMs ]
      properties:
        challengeId:
          type: string
          minLength: 1
          maxLength: 128
        challenge:
          type: string
          minLength: 1
        expiresAtMs:
          type: integer
          minimum: 1
    MobileSessionAttestation:
      type: object
      additionalProperties: false
      required: [ type, token ]
      properties:
        type:
          type: string
          enum: [ play_integrity, app_attest ]
        token:
          type: string
          minLength: 1
    MobileSessionAttestRequest:
      type: object
      additionalProperties: false
      required: [ challengeId, platform, deviceIdHash, attestation ]
      properties:
        challengeId:
          type: string
          minLength: 1
          maxLength: 128
        platform:
          type: string
          enum: [ android, ios ]
        deviceIdHash:
          type: string
          minLength: 8
          maxLength: 256
        attestation:
          $ref: '#/components/schemas/MobileSessionAttestation'
    MobileSessionAttestResponse:
      type: object
      additionalProperties: false
      required: [ accessToken, expiresAtMs, integrity ]
      properties:
        accessToken:
          type: string
          minLength: 1
        expiresAtMs:
          type: integer
          minimum: 1
        integrity:
          type: string
          enum: [ strong, device, basic, unknown ]
    PushTokenPayload:
      type: object
      additionalProperties: false
      required: [ token, deviceId, platform, provider, appVersion, locale, timezone ]
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 512
        deviceId:
          type: string
          minLength: 1
          maxLength: 256
        platform:
          type: string
          enum: [ ios, android ]
        provider:
          type: string
          enum: [ apns, fcm ]
        appVersion:
          type: string
          minLength: 1
          maxLength: 64
        locale:
          type: string
          enum: [ fr, en ]
        timezone:
          type: string
          minLength: 1
          maxLength: 128
...